Improve validation and exception handling

+++ src/main/java/com/victor/mc_cd_catalog/catalogs/infrastructure/web/models/CatalogResponse.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Builder
 public class CatalogResponse {
     int page;
-    int productCount;
+    long productCount;
     List<CatalogProductDto> catalogProducts;
 }
